pyproject.toml
[project]
name = "heterogeneity_code"
authors = [{name = "..."}]
version = "0.1.0"
description = "Short description of the project."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "umap-learn>=0.5",
    "pysfo @ file:///storage/Dropbox/01_main/dev/pysfo",
    "clean_nport @ file:///storage/Dropbox/01_main/dev/clean_nport",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
